+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## not released
 
 - Fix: #45 Error message when deleting old ZIP backupsets
+- Disable Joplin DirectoryPath selector and use string input again because of several bugs [#6692](https://github.com/laurent22/joplin/issues/6692) and [#6693](https://github.com/laurent22/joplin/issues/6693)
 
 ## v1.1.0 (2022-07-11)